Perform early stack overflow checks before writing ActRec
Summary:
Move the early stack overflow check before writing the callee's ActRec, a step
towards deferring ActRec write to the function entry.

This diff refactors some of the stack overflow routines to not depend on the
callee frame being spilled. This will be useful by the next diff that will
allow FCall with unpack to always use prologues by repacking arguments before
the prologue is reached, which requires a stack overflow check.

It also brings the behavior closer to the interpreter, which performs a stack
overflow check before doFCall().

Random notes:
- we just throw FatalErrorException instead of using raise_error() which
  reduces to the same to avoid potential reentrancy issues if raise_error()
  gets modified
- fcallHelper() now spills missing NullUninits, making it safe to throw before
  spilling ActRec and making it safe to throw from doFCall() for inout methods
  (this was broken, but very unlikely to happen)

Reviewed By: ricklavoie

Differential Revision: D18826997

fbshipit-source-id: 90409c97ccdfda180c8972945501474bb3cb7cb5