Use ppx_deriving ord to derive some comparisons in Typing_defs
Summary: There's really no reason to manually write out ordinal numbers for variants when we have PPX. In this diff I've set up a few that were easy to migrate. The derived functions take the form `val compare_type_name : type_name -> type_name -> int`. We can't yet derive over `'a ty` because it's a GADT, but there is work underway to remove the GADT so we can derive there too.

Reviewed By: kmeht

Differential Revision: D18718949

fbshipit-source-id: 305aa81b0672b1a3f92e5080d0dafadcf58cf0f5