Add basic support for function reactivity attributes
Summary:
The definition of the target function / method is extracted including its attributes, in particular, the attributes that determine the function reactivity.  The stubs of the dependency functions / methods are generated from their type and are currently lacking reactivity attributes.  This may lead to type errors in the extracted code.  Consider, for example:
```
<<__Rx>>
function reactive(): void {}

<<__Rx>>
function call_reactive(): void {
  reactive();
}
```
Extraction of `call_reactive` produces
```
function reactive(): void {}

<<__Rx>>
function call_reactive(): void {
  reactive();
}
```
which is not well-typed because `call_reactive` is marked as reactive but calls `reactive`, which doesn't have reactivity annotations and hence is considered non-reactive.

The mapping between the user attributes attached to a function (and its parameters) and the internal representation of reactivity is somewhat non-trivial.  Here I'm adding very basic support for emitting reactivity attributes for functional dependencies.

Reviewed By: kmeht

Differential Revision: D18783619

fbshipit-source-id: 81ed702894dad4226a61b1519dd22d1f844a0a9d