Support arbitrary tv-layout for PropInitVec
Summary:
We want both objects and PropInitVecs to have the same layout as
there's a number of nice optimizations that rely on this fact. Now we get to
restore some of these.

There's at least one problem with this, as PropInitVec uses a TypedValueAux to
store a "deep init" bit on each property, which marks whether or not it needs
special initialization. There's no space in tv_layout for extra data of this
kind and I don't want to provide for it.

So I solve that problem by shoving the bits in a buffer after the tv_layout.
This ought to be fine since it's only accessed (after the class is loaded) from
deepInitHelper which is already shoved onto a slow path.

It's also worth noting that this changes the layout of PropInitVec when wide mode is _off_ since it now stores the DeepInit bits separately.... eventually it might be worth adding some of the functionality of TypedValueAux back to TvLayout but it seems like a relatively minor tweak

Differential Revision: D18646826

fbshipit-source-id: 370dc6894488b13a0e2c8ca120cce996d5d58cbe