Optimize 7-up heap scan and release.
Summary:
The observation was that both of these operations got much slower when
switching to 7-up layouts, so now we get to stop apologizing for the weirdness
of the layout and start exploiting it.

We can avoid an awkward iteration step and also several fetches if we iterate
over the container in chunks instead of one value at a time. This also allows
us to fetch the types in one go instead of one-at-a-time.

Another tweka to make this more practical is to assume that the last chunk of
the container (which might have a number of nonsense types in the last type
word) has any extra space filled with 0s (KindOfUninits) so that we really
don't have to muck with deailing with the remainder mod 7. This means it's also
advantageous to store the countablePropsEnd of the class as a quick_index.

Differential Revision: D18647108

fbshipit-source-id: 1bb5705262a6078de6fe555ebc5969236f18b471