Add "quick indexes" to tv_layout.
Summary:
So, dividing by seven isn't particularly problematic (especially when
the index is constrained to a 16-bit number) but it does present issues of code
size and ideally we don't want to be doing it in hot code. (this is especially
important in hotter runtime code)

This adds a notion of a 'quick index' to tv_layotu which is an index that might
take effort to precompute but which provides fast access into the container. In
the extreme case, you might imagine this being just a pair of offsets to the
type and value byte.

In the case of 7-up this is just a single 16-bit number where the lower 3 bits
denote the number's remainder mod 7 and the remaining bits represent the
quotient mod 7. This happens to be ideal since we can compute (index / 7) * 8
just by masking.

Differential Revision: D18647107

fbshipit-source-id: 69d7ac7f13e740a630a2208fb07af070433d576f