Uniquify symbols based on flags
Summary:
This diff makes it so that symbols are unique with respect to flags which means that if we see the same symbol in multiple CUs or TUs with the same configuration of flags and some heuristic, then we only register the first observed entry.

The heuristic is provided in the code.

GDB appears to do this optimization to speed up symbol lookup.

Reviewed By: markw65

Differential Revision: D18837154

fbshipit-source-id: b5d399ee3d1eae6a15f6a6fcd34702eb05227426