Improve error messages and clean up subtyping for destructuring
Summary: This the result of factoring {D18633271} into a more manageable diff. It splits destructuring into two main categories: list-like and tuple-like. Tuples and pairs have many types to be split into the destructure boxes, while lists are monomorphic. I removed the default subtype `invalid ()` cases in favor of the explicit arity errors and a new error that says a type cannot be destructured.

Reviewed By: Wilfred

Differential Revision: D18693728

fbshipit-source-id: 1f6be56463b397439eab0de2e4ee4ec1c0c5ad1c