exit_on_parent mandatory
Summary:
There's an optional flag `exit_on_parent_exit`. But everyone passes it. I'm going to make it mandatory.

(I think the `exit_on_parent_exit` mechanism is a valuable one, and everyone should be using it, and I don't want to see new things come along which fail to use it.)

Reviewed By: arxanas

Differential Revision: D18754910

fbshipit-source-id: 973bfb3fa52cd775fdc5f9326bd8c97e59352dcb