sIDE can log tracking_id
Summary:
I'm working on a tracking id, so I can track how a message is handled from the moment it arrives in clientLsp, through its handling by clientIdeDaemon, back to the response to the client.

The reason for that is (1) gain insight into places where sIDE is crashing, (2) gain insight into why sIDE sometimes takes so long after it's initialized before it successfully handles its first message. I have some hunches, but I ultimately want to see data for exactly how many typechecks it's doing in order to answere that, how many decl-fetches, and why. And I want to see how many people are affected, and how much. Ted had sometimes observed the initial delay was related to typechecking time for a file so I'll look into that too. I have a hunch there might be stuff piling up in an sIDE queue, like there sometimes is with hh_server, and I want to get to the bottom of that too.

This diff plumbs the tracking_id into clientIdeDaemon.

We don't yet do anything with it yet, on the simple grounds that clientIdeDaemon doesn't report any telemetry on message-handling via HackEventLogger. But once it does, we'll pass in the tracking id.

Reviewed By: arxanas

Differential Revision: D18806087

fbshipit-source-id: 3023437844aacee833c8bc0a335ebc8668881695