relate sIDE id to that of client
Summary:
My goal is to gather detailed telemetry for sIDE.

The way things work in hh_server is that the master process gets an "init_id", a unique per-process random string, and it initializes its telemetry-logging with that init_id.

Each worker process calculates its own "worker_id". And it initializes its telemetry-logging with init_id being `master_id ^ "." ^ worker_id`. In this way you can recognize distinct workers in the logs, and you can recognize which master process they're associated with.

I'm just doing the same for the sIDE daemon.

We could do this differently, by correlating telemetry based on hostname and username and time. But having the IDs direct in the telemetry seems more straightforward, and is how things are done elsewhere in our codebase.

My ultimate goal in this diff stack is to be able to relate telemetry that comes from clientLsp, to telemetry that comes from clientIdeDaemon. My goal is to be able to relate them conveniently in two ways: (1) have a "init_id" relation between telemetry logs from the two process instances, per this diff; (2) have a per-message "tracking_id" relation also.

The reason for that is (1) gain insight into places where sIDE is crashing, (2) gain insight into why sIDE sometimes takes so long after it's initialized before it successfully handles its first message. I have some hunches, but I ultimately want to see data for exactly how many typechecks it's doing in order to answere that, how many decl-fetches, and why. And I want to see how many people are affected, and how much. Ted had sometimes observed the initial delay was related to typechecking time for a file so I'll look into that too. I have a hunch there might be stuff piling up in an sIDE queue, like there sometimes is with hh_server, and I want to get to the bottom of that too.

Reviewed By: arxanas

Differential Revision: D18766033

fbshipit-source-id: 27e9121f3dede5d3083cf36c3794b0006e53ef19