Remov funky isReturning logic from backtrace.cpp
Summary: In backtrace.cpp, when we're deciding whether to include locals, this, and frame metadata in the backtrace, we check fp->localsDecRefd() but also for some reason check for 5 specific bytecodes: the 3 Ret flavors but also inexplicably CreateCont and Await (but not AwaitAll, Yield, or YieldK: the other bytecodes that may suspend). We don't need this, we should be able to rely solely on fp->localsDecRefd().

Reviewed By: ricklavoie

Differential Revision: D18874723

fbshipit-source-id: ee42dff7cac4328d75fc52fab84d1ba746b47f04