Check for a SrcRec after acquiring the code lock
Summary:
We check if there's already a SrcRec, acquire the lock, generate some
code, and then assert there's no SrcRec. But its quite possible that
another thread created it while we were getting the lock.

The docs say that the SrcDB is protected by the metadata lock, and
indeed we do acquire it before inserting into it. But since this is
the *only* place that we insert into the SrcDB, its sufficient to
check after grabbing the code lock.

Reviewed By: ottoni

Differential Revision: D18869885

fbshipit-source-id: 3e2b13bf0b5aeefdd09c42a9436d95b84efde9ed