Restructure prologue emission (take 2)
Summary:
Currently, emitting a prologue requires holding the code and metadata locks for the whole operation, but they're only required for part of it.

Restructure the code a bit, and allow callers to pass in a CodeMetaLock to just lock for the critical section. A later diff in the stack will add uses of the CodeMetaLock parameter.

Reviewed By: binliu19

Differential Revision: D18843946

fbshipit-source-id: c9ad4cea8ae4ba98100d6fc318053e44007c4ddb