Check arity checks on class consts
Summary:
It turns out that we are not checking type arity on class consts either.

We should refactor `type_params_arity_check` so that it is opt out and not opt in. And probably write a wiki to warn against the pattern as well.

This has likely regressed at some point... Perhaps when we were migrating the check over to use the tast?

Reviewed By: kmeht, Wilfred

Differential Revision: D18257361

fbshipit-source-id: c69bc37edfe88494f73c6cdc5781578ef819b570