Initialize async PC in backtraces
Summary:
We started crashing in createBacktrace recently, due to decoding an invalid op at the backtrace PC. These crashes only seem to occur when we create a backtrace due to a stack overflow. I ran HHVM at some older revisions (specifically, at release S) and found that we still crashed this way, indicating that the bug has been present for at least two weeks.

One recent diff that changes backtrace PC computation is D18125503. In this diff, in one of the async cases, we set the  PC to 0 (a valid offset value, but not one that we'd expect to be in the function) rather than func->base(). That may be the bug - I'm not sure.

Reviewed By: markw65

Differential Revision: D18751926

fbshipit-source-id: 7a5d19ded84f5d2f82d672b8c68ba6fc0337900f