Stricter verification for iterator exceptions
Summary:
The verifier doesn't check iterator invariants as thoroughly as it should. It's possible for some ops to be unguarded by iterators, and it's possible for us to have live iterators when we leave a function. We'd like to prevent such cases, and to make it possible to replace IterBreak with a sequence of IterFrees + a Jmp. Validating iterator exception blocks is key.

Prior to this change, the exception logic in check-func was: "push state from the first non-SilenceEnd bytecode in this block to the exn block". This logic can miss some cases where iterators would be freed twice. In particular, if there's an IterFree followed by an instruction that would throw, it would push the pre-IterFree state to the exn edge and miss the fact that the iterator could be freed twice.

The new logic says the following:
- Any mayTakeExnEdges instruction with live iterators must be guarded by an exn.
- When we encounter the first mayTakeExnEdges instruction in a block, we push that state to the exn block.
- We won't process any blocks that haven't gotten their state initialized.

The last rule is required for when a block has no mayTakeExnEdges instructions at all. For such blocks, we used to have some special-case logic around Silence::End opcodes to try to still initialize the state of the unreachable exn block. If we wanted to fake the state in this way for iterators, we'd have to fake it differently than we do for Silence::End. Rather than faking the state for an unreachable exn block, we avoid verifying such constraints. (Note that if there's any other path to the exn edge for a loop, e.g. through a block other than the iter break, we will verify its state.)

Fatal and Jmp are two cases where it's not obvious that we can return false for mayTakeExnEdges. For Fatal, the justification is that we're failing the request anyway; we don't clean up fatals using exception blocks. For Jmp, we have to consider the handle-surprise case, which can re-enter. We have try-catch guards around these callbacks, so even if one throws an error, it shouldn't result in us taking the exn edge within the function being verified.

With this new logic, we can also re-enable the iterator type consistency checks, which are much stricter than existing iterator checks. (Without these checks, all we ensure is that the first RPO path to a iterator exn block is consistent with its expected iterator state.)

Reviewed By: ricklavoie

Differential Revision: D18525598

fbshipit-source-id: f7a7aed7d618e018c9de2ccbf4e83942adadab9d