Insert global results constructed from autocomplete re-rank server response
Summary:
For global name completion, the ranked autocomplete server will respond with new completion items that are not originally from the completion list queried from Glean. In this case, we want to construct new completion result based on the symbol name.

This diff also tries to resolve the incorrect docblock for ranked completion items. Previously, the Completion.kind field is set to `Completion.Event` for ranked items (in order to have the small lightning icon), but the docblock resolver is dependent on the value of the completion kind and try to locate the symbol: https://our.intern.facebook.com/intern/diffusion/FBS/browse/master/fbcode/hphp/hack/src/server/serverDocblockAt.ml?lines=164-194

To fix this, we pass the completion kind value into the `data` field.

Differential Revision: D18861700

fbshipit-source-id: 1f97b137335e7b4c5d5de1d82364e80c67c15c44