last strongly typed
Summary:
My goal is to move away from clientLsp using ad-hoc parsing of the jsonrpc message, to having a single canonical parse of that jsonrpc into a strongly-typed ocaml ADT.

This diff is the last bit I care to do for now. The thing is, there are still a few properties hanging off the jsonrpc (timestamp, tracking_id) which we need to use.

I decided to call these things "metadata" about the request. In effect, we have to have both (1) strongly-typed ocaml ADT, (2) additional metadata about the request that's not part of the jsonrpc payload in and of itself.

A future diff will probably want to continue this refactoring - D18876626. Once we finish that, we'll have a strongly typed guarantee that no part of the code is digging into the raw jsonrpc.

Reviewed By: arxanas

Differential Revision: D18875009

fbshipit-source-id: da3238673a8f14a44430b229f98e17c9f5264593