get_uri_opt
Summary:
In the olden days, prior to LSP, the editor used to invoke `hh_client --type-at-pos <filename> <line> <col>` or similar. We could look in our telemetry at the command-line and figure out what filename was being operated upon.

When LSP started to roll out, we recorded in telemetry the Jsonrpc request and response (truncated in case it was excessively long). This also gave the ability to figure out what filename was being operated upon.

But all that jsonrpc quickly became too much, and we had to stop collecting it.

I want to restore the ability to see which filename we're operating upon. This diff, given any LSP requests, returns the filename associated with it (if there was one).

Reviewed By: arxanas

Differential Revision: D18875559

fbshipit-source-id: 4314aa5d73ac0cccd3fac422c9e25c830313331d