add types for $test/startHhServer
Summary:
My goal is a single canonical place where we parse jsonrpc into an ADT. Formerly only Flow did this; now Hack will do it as well. That ADT is `Lsp.lsp_message`.

There are a few custom LSP requests we invented for sake of testing. This diff adds them to the ADT.

Reviewed By: arxanas

Differential Revision: D18874997

fbshipit-source-id: 137d0f75253983e8ee3be67fc4bbd2f83ac193a8