clientLsp logs filename
Summary:
In the olden days, prior to LSP, the editor used to invoke hh_client --type-at-pos <filename> <line> <col> or similar. We could look in our telemetry at the command-line and figure out what filename was being operated upon.

When LSP started to roll out, we recorded in telemetry the Jsonrpc request and response (truncated in case it was excessively long). This also gave the ability to figure out what filename was being operated upon.

But all that jsonrpc quickly became too much, and we had to stop collecting it.

This diff reestores the ability, in our telemetry, to see which filename we're operating upon.

Reviewed By: arxanas

Differential Revision: D18875713

fbshipit-source-id: 662e7c54186526ffb9f5ad2fe6a8b36f4b0cdcb4