add types for ToggleTypeCoverage
Summary:
My goal is a single canonical place where we parse jsonrpc into an ADT. Formerly only Flow did this; now Hack will do it as well.

That ADT is `Lsp.lsp_message`. Flow never handled ToggleTypeCoverage, so we never noticed that the ADT lacked a constructor for it. This diff adds the constructor.

Reviewed By: arxanas

Differential Revision: D18874996

fbshipit-source-id: 4518884383fa7e438b939df7548dfc45a91c057d