store request with callback handlers
Summary:
My goal is to use a single canonical `Lsp_fmt.parse_lsp` function to parse the incoming jsonrpc, and then the remainder of clientLsp can operate off a strictly-typed ocaml ADT which represents the requests.

One hiccup is that you can't parse an jsonrpc response in isolation. The only way to parse it is if you know what method it's in response to. Indeed one of the parameters to `parse_lsp` is a callback which is called in case the jsonrpc is a response and asks "what is the original requests associated with this response id?"

We weren't previously stored the request associated with an outstanding callback id. This diff adds it.

Reviewed By: arxanas

Differential Revision: D18874987

fbshipit-source-id: 745805ae6682a9fddb7fd198996d4b1f31a1505e