strongly type handle_client_message
Summary:
My goal is to have a single canonical parse of jsonrpc into a strongly-typed ocaml ADT, and then have the rest of the code in clientLsp consume this ADT.

This diff is an incremental step on the way. It switches over the bulk of `handle_client_message` to use the strongly-typed ADT.

One other change. It used to be that each handle_client_message case for serverless-IDE looked like this:
```
| ( ( In_init { In_init_env.editor_open_files; _ }
    | Main_loop { Main_env.editor_open_files; _ }
    | Lost_server { Lost_env.editor_open_files; _ } ),
    c )
    when env.use_serverless_ide && c.method_ = "textDocument/completion" ->
```

I've switched over to obtaining `editor_open_files` at the start of `handle_client_message`, so the case now looks like this:
```
| (_, RequestMessage (_id, CompletionRequest params))
  when env.use_serverless_ide ->
```

There are only five states (Pre_init, Post_shutdown, In_init, Main_loop, Lost_server). The first two are already handled by earlier match cases. That's why this change has no behavioral effect.

We lose a tiny little bit in terms of type-safety, but I think we more than gain back enough in terms of readability of each match case.

Reviewed By: arxanas

Differential Revision: D18875004

fbshipit-source-id: d256e37df2c8dacdf4f07183853e29328f0a40dd