no longer log LSP json
Summary:
We initially thought we could log a truncated form of each jsonrpc message between client and server in our telemetry.

It became quickly apparent that the volume of data would be too vast. We stopped collecting this even before LSP had reached 100% rollout.

This diff is just trivial cleanup to remove the long-dead paths.

Reviewed By: arxanas

Differential Revision: D18874999

fbshipit-source-id: 157b8a7f18a020827837cfbcb4a2d7414bf58947