add type for SignatureHelp
Summary:
My goal is a single canonical place where we parse jsonrpc into an ADT. Formerly only Flow did this; now Hack will do it as well.

That ADT is `Lsp.lsp_message`. Flow never handled SignatureHelp, so we never noticed that the ADT lacked a constructor for it. This diff adds the constructor.

Reviewed By: arxanas

Differential Revision: D18874991

fbshipit-source-id: 07f0d8423298010f114ed0255227b864271d0565