add types for CompletionItemResolve
Summary:
My goal is a single canonical place where we parse jsonrpc into an ADT. Formerly only Flow did this; now Hack will do it as well.

That ADT is `Lsp.lsp_message`. Flow never handled CompletionItemResolve, so we never noticed that the ADT lacked a constructor for it. This diff adds the constructor.

Reviewed By: arxanas

Differential Revision: D18874994

fbshipit-source-id: 15280bed7f892cda3e94b5533d524631ba72d331