add textDocument/typeDefinition
Summary:
My goal is a single canonical place where we parse jsonrpc into an ADT. Formerly only Flow did this; now Hack will do it as well.

That ADT is Lsp.lsp_message. Flow never handled TypeDefinition, so we never noticed that `Lsp_fmt.parse` missed it. This diff adds it in.

Reviewed By: arxanas

Differential Revision: D18875509

fbshipit-source-id: 429e1e4aec513be03f082eee1d3f5e59aaee0eaf