add types for GoToImplementation
Summary:
My goal is a single canonical place where we parse jsonrpc into an ADT. Formerly only Flow did this; now Hack will do it as well.

That ADT is `Lsp.lsp_message`. Flow never handled GoToImplementation, so we never noticed that the ADT lacked a constructor for it. This diff adds the constructor.

Reviewed By: arxanas

Differential Revision: D18874993

fbshipit-source-id: aa50137be78afcd015eec6d70ed255f076dac2be