When adding a dependency on the class constructor, also add a dependency on its signature
Summary: When adding a dependency on a class property, we also add a dependency on the class constructor (because the property may need to be initialised in the constructor), but at the moment we don't add dependencies on the types occurring in the constructor signature.  See the example below in which this manifests and causes the extracted code to fail to type check with an "Unbound name" error.

Reviewed By: kmeht

Differential Revision: D18851137

fbshipit-source-id: 8fefd3e53be4a3b210c0f660a825db53e7846317