Add implementations of methods required by a parent class or interface
Summary:
Consider the following (almost trivial!) example:
```
abstract class BB {
  abstract public function f(): void;
}

interface II {
  public function g(): void;
  public function h(): void;
}

class CC extends BB implements II {
  public function f(): void {}
  public function g(): void {}
  public function h(): void {}
}

function with_implementations(BB $b, II $i, CC $c): void {
  $b->f();
  $i->g();
}
```
If you ask the extraction tool to extract `with_implementations`, you will get
```
<?hh
function with_implementations(BB $b, II $i, CC $c): void {
  $b->f();
  $i->g();
}
class CC extends \BB implements \II {}
abstract class BB {
  abstract public function f(): void;
}
interface II {
  public function g(): void;
}
function extract_standalone_make_default(): nothing {
  throw new Exception();
}
```
Huh?  The class `CC` will be extracted with its parents `BB` and `II` but no stubs for the methods `f` and `g` required to satisfy the parent class and interface signatures will be generated in `CC`.  The extracted code produces a type error.

The problem is that at the moment we only add dependencies on methods required by built-in interfaces.  We could add dependencies on all methods required by parent interfaces (and all abstract methods required by the parent class), but that would mean we would generate a lot of redundant code.  (E.g., above we don't want to generate a stub for `h` because it is never called on objects of type `II` and hence doesn't need to be pulled into the extracted version of `II`, and therefore the extracted version of `CC` doesn't need to have this method defined.  We would ideally want to pull in as few dependencies as possible.  On the other hand, in order to add a dependency e.g. on `C::g` above we need to first figure out that `I::g` is a dependency.  In other words, when adding a dependency on an interface method, we also need to add dependencies on implementations of this method in all implementors of the interface.  Unfortunately, this information is not readily available, so I wrote a function that goes over the set of dependencies collected so far, and for each class dependency iterates over its parents and collects dependencies on the methods in this class required by the parents, and then adds these dependencies to the set.  I believe this needs to be iterated to a fixed point because adding a dependency on an implementation may pull in new types which may require pulling in more implementation dependencies.

Reviewed By: andrewjkennedy

Differential Revision: D18832490

fbshipit-source-id: fd940f58ed76fd1fbb4b85d010456de7f9409725