Fix a bug with adding type constant dependencies
Summary: The function `add_dep` is written using a type visitor instead of explicit recursion, and when adding a dependency on `C::Tx` it calls `this#on_type` on the type `Tx` is bound to, which may be defined in terms of another type constant, say `const type Tx = this::Ty::Tx;`, so that `this#on_type` is called on `this::Ty::Tx`.  What this means is that depending on where the dependency on `C::Tx` is added, the `~cls` argument to `add_dep` that is shared by all visitor methods may not be set to the right class and `this` in `this::Ty::Tx` will be resolved to the wrong class, causing `DependencyNotFound` exception to be thrown.  (See the new test case added below.)

Reviewed By: vsiles

Differential Revision: D18808487

fbshipit-source-id: f0e1aa82c7cb7a574fe9cd87e40aa294685cfac9