Embed the provider backend in the provider context
Summary:
This information all logically belongs together. It was separate because I didn't have a clear idea of what the provider context entailed when I made the abstraction.

For now, we just get the global provider context whenever we make a new provider context. In the future, it should be passed as a parameter. (It's likely that we'll have an `empty_from_server_env` constructor for `Provider_context.t`.)

Reviewed By: ljw1004

Differential Revision: D18768763

fbshipit-source-id: 678a607f200aeb14370803246f85a66264642084