Remove typecheck_xhp_cvars option
Summary:
I don't recall the historical context as to why we didn't enforce tparam arity in XHP classes, but moving forward we should always enforce it. This removes the `typecheck_xhp_cvars` option, making the enforcement mandatory.

Someone's guess on historical context: " I think it was the HHVM parser for XHP didn’t support generics so we allowed them to be without attributes. This was supposed to be a 'short term fix' that never got cleaned up"

Reviewed By: Wilfred

Differential Revision: D18257005

fbshipit-source-id: 6ee1462d54c039e354757f1b5a77250fde3f036d