avoid potential stop race
Summary:
Prior to this diff, we'd happily send a 'shutdown' message to the daemon, leave the state as "Initialized" in the meantime, and then other top-level message loops would be able to act based on the state being Initialized. To be precise: although they think the state is "Initialized", they'd be interacting with a daemon who has already received the shutdown message.

I don't know if that was a problem, and I have no clear evidence of it having been a problem.

But it still feels dodgy. This diff therefore changes the order, so state is marked as "Stopped" immediately before we send the shutdown message to the daemon.

Reviewed By: arxanas

Differential Revision: D18911655

fbshipit-source-id: ebc367baa3e193a5c1e93a51415038580edc8068