Eliminate IterBreak instructions
Summary:
In hackc, rather than emitting "IterBreak L0 <(Iter) 1, (Iter) 0>", we can emit "IterFree 1; IterFree 0; Jmp L0". With the verifier changes in this stack, this change should be relatively safe.

Once we've made this change, we can delete this instruction and the IterTable immediate type from HHVM. Most of the lines in this change are due to deleting the immediate.

Reviewed By: ricklavoie

Differential Revision: D18465805

fbshipit-source-id: 4c6b6ba90f7d88dfc687ef4992c063bf9c33b836