Create `get_global_context_or_empty_FOR_MIGRATION` constructor for `Provider_context`
Summary: This constructor indicates that the requisite information should have been passed in as parameters, but wasn't. We'll just get the global information for now. As part of refactoring, we'll try to push calls to `get_global_context_or_empty_FOR_MIGRATION` further up until we reach a point where the required information is available.

Reviewed By: ljw1004

Differential Revision: D18867183

fbshipit-source-id: 4b5fb50068370cfa9effe15527f792a2cf25f2f0