make compute_tast return a record
Summary:
My goal is to be able to track telemetry about typechecking performance and decl-fetches. As part of this, I'll want compute_tast to return richer information about how it performed (not just the tast+errors results).

This is a trivial diff that makes compute_tast return a record, in preparation for adding new fields to it.

Reviewed By: 2BitSalute

Differential Revision: D18891662

fbshipit-source-id: f79b23a2e77b006cc6ada87342075a9c542d925f