expose defered_decl count
Summary:
The deferred-decl machinery already counts how many decls needed to be fetched. But it only incremeents that count if a threshold is provided (and hence has the potential to raise a Defer exception).

I want to re-use that counting machinery, for use in serverless-IDE and other uses of `compute_tast_and_errors`. But those uses are typically single-threaded, don't have a worker queue, and can never defer.

This diff therefore (1) exposes the counter, (2) lets the counter be incremented even if no threshold is provided.

Reviewed By: 2BitSalute

Differential Revision: D18879091

fbshipit-source-id: 5ad7b65bf857de69bbd4c7edfb92e2f672201bc2