hh_shared.c: optionally reset the dependency graph in `hh_save_dep_table_blob`
Summary:
While testing something else, I found that I forgot to reset the dep table after saving the blob. Adding this functionality, and adding a test for it.

I also noticed that the ~replace_state_after_saving is not the correct name for the flag to saving the dependency graph as a blob. The necessary flag is ~reset_state_after_saving.

Renaming the flag, I realized that it's being used in the SaveStateService module, but that code path is not live, so I removed the dep-table-as-blob option from that module altogether.

Reviewed By: ljw1004

Differential Revision: D18899522

fbshipit-source-id: d6d1cc625c66dfb139f97f13fcdcf6698157dd97