Make sqlite able to search for symbols within a child namespace
Summary:
As reported by Fred Emmott some users would like workspace symbol search to be able to find symbols that are within non-root namespaces.

For example, if the user does a workspace symbol search for "foo", they would like to be able to find all of the following:
* \Foo
* \Foobar
* \Some\Path\Foo

This diff renames the "ac_no_namespace" context to specify that it is used for workspace symbol searching.  It then modifies the behavior of sqlite for workspace symbol search to include a search for these items.

Reviewed By: fredemmott

Differential Revision: D18897635

fbshipit-source-id: c0f795110170a880cc5e59b785f0d2387052bf0d