Speed-up determine spilling needed logic in vasm-graph-color
Summary:
Speed-up the logic in vasm-graph-color which determines if spilling is
needed in a block. Do a very cheap coarse check at the beginning of
the block. If the sum of the live-in Vregs and Vregs defined in the
block is below the size of both reg classes, the block cannot cause a
spill, so we can skip it entirely. Furthermore, if we are processing
the block, as soon as we detect a spill, we can move onto the next
block, and not finish processing the current block.

Reviewed By: mofarrell

Differential Revision: D18867754

fbshipit-source-id: eddf1a86db1f80a195b06023c3429d0a3948fc11